#585: dedup_test.go was 3 lines over the CI ceiling
Found by running the file-size gate on EVERY branch after GPT caught the same
defect on the #529 branch. The reviewer pointed at one instance; the question
worth asking was whether I had done it elsewhere. I had.

internal/watch/dedup_test.go reached 803 lines against the repo enforced 800,
so `make repo-hygiene` fails and CI fails with it. My newCappedStore helper and
its guard test pushed it over.

Moved them to retention_cap_test.go rather than shaving lines: they are about
RETENTION CONFIGURATION, not deduplication, so they belong with the other
retention tests and dedup_test.go goes back to being about dedup. 803 -> 750.

The move script asserts on the head and tail of the block it lifts and refuses
to write if either is unexpected -- which caught an off-by-one on the first run
of the equivalent script for #529, before it could corrupt anything.

File-size gate passes with a true exit 0; internal/watch passes.
